fixedRelaxation algorithm working
