+++ coupling_algorithms/algo_fixedRelaxation.h
@@ -50,6 +50,7 @@
 
 #include "../util.h"
 #include "../config.h"
+#include <iterator>
 
 namespace mui {
 
@@ -61,33 +62,221 @@ public:
 	using point_type = typename CONFIG::point_type;
 
 	algo_fixedRelaxation( REAL undRelxFac = 1.0, 
- 	  std::vector<std::pair<point_type, REAL>> ptsVluInit = std::vector<std::pair<point_type, REAL>>() ):
- 	  initUndRelxFac_(undRelxFac),
- 	  ptsVluPositive_(ptsVluInit) {
+ 	  std::vector<std::pair<point_type, REAL>> ptsVluInit = 
+		std::vector<std::pair<point_type, REAL>>() ):
+ 	  initUndRelxFac_(undRelxFac) {
 
-	    posNegCounter_ = 1;
-	    ptsVluNegative_= std::vector<std::pair<point_type, REAL>>();
+		undRelxFac_ = initUndRelxFac_;
 
+		if (!ptsVluInit.empty()) {
+			ptsTimeVlu_.push_back(
+				std::make_pair(
+				std::numeric_limits<time_type>::lowest(),ptsVluInit
+				)
+			);
+		}
 	}
- 
+
 	//- relaxation based on single time value
 	template<typename OTYPE>
-	OTYPE relaxation(point_type focus, const OTYPE &filteredValue) const {
+	OTYPE relaxation(time_type t, point_type focus, OTYPE filteredValue) const {
 
-		posNegCounter_ *= -1;
+		OTYPE filteredOldValue = 0.0;
 
-		OTYPE relaxedValue = filteredValue;
+		if (ptsTimeVlu_.empty()) {
+			
+			filteredOldValue = 0.0;
 
-		return relaxedValue;
+			std::vector<std::pair<point_type, REAL>> ptsVluTemp{
+				std::make_pair(focus,get_relaxed_value(filteredValue,filteredOldValue))
+				};
+
+			ptsTimeVlu_.push_back(std::make_pair(t,ptsVluTemp));
+
+			return get_relaxed_value(filteredValue,filteredOldValue);
+
+		} else { // ptsTimeVlu_ not empty
+
+			auto presentIter = std::find_if(ptsTimeVlu_.rbegin(), ptsTimeVlu_.rend(), 
+				[t](std::pair<time_type,std::vector<std::pair<point_type, REAL>>> b) {
+					return (t - b.first) < std::numeric_limits<REAL>::epsilon();
+				});
+
+			auto previousIter = std::find_if(ptsTimeVlu_.rbegin(), ptsTimeVlu_.rend(), 
+				[t](std::pair<time_type,std::vector<std::pair<point_type, REAL>>> b) {
+					return b.first < t;
+				});
+
+			if ((presentIter == std::rend(ptsTimeVlu_)) && 
+				(previousIter == std::rend(ptsTimeVlu_)) ) {
+
+				std::cerr << "Non-monotonic time marching does not (yet) supported for the Fixed Relaxation coupling method! " << std::endl;
+
+			} else if ((presentIter != std::rend(ptsTimeVlu_)) && 
+				(previousIter == std::rend(ptsTimeVlu_)) ) {
+
+					auto ptsRelxValIter = std::find_if(presentIter->second.begin(), 
+						presentIter->second.end(), [focus](std::pair<point_type, REAL> b) {
+					return normsq(focus - b.first) < std::numeric_limits<REAL>::epsilon();
+					});
+
+					if ( ptsRelxValIter == std::end(presentIter->second) ) {
+
+						// Interpolate the relaxed value by N2_linear
+						REAL r2min_1st = std::numeric_limits<REAL>::max();
+						REAL r2min_2nd = std::numeric_limits<REAL>::max();
+						OTYPE value_1st = 0, value_2nd = 0;
+						for( size_t i = 0 ; i < presentIter->second.size() ; i++ ) {
+							REAL dr2 = normsq( focus - presentIter->second[i].first );
+							if ( dr2 < r2min_1st ) {
+								r2min_2nd = r2min_1st;
+								value_2nd = value_1st;
+								r2min_1st = dr2;
+								value_1st = presentIter->second[i].second ;
+							} else if ( dr2 < r2min_2nd ) {
+								r2min_2nd = dr2;
+								value_2nd = presentIter->second[i].second ;
+							}
+						}
+
+						REAL r1 = std::sqrt( r2min_1st );
+						REAL r2 = std::sqrt( r2min_2nd );
+
+						auto relaxedValueTemp = ( value_1st * r2 + value_2nd * r1 ) / ( r1 + r2 );
+
+						presentIter->second.push_back(std::make_pair(focus, relaxedValueTemp));
+
+						return relaxedValueTemp;		
+
+					} else {
+
+						return ptsRelxValIter->second;
+
+					}
+
+			} else if ((presentIter == std::rend(ptsTimeVlu_)) && 
+				(previousIter != std::rend(ptsTimeVlu_)) ) {
+
+					auto ptsRelxValIter = std::find_if(previousIter->second.begin(), 
+						previousIter->second.end(), [focus](std::pair<point_type, REAL> b) {
+					return normsq(focus - b.first) < std::numeric_limits<REAL>::epsilon();
+					});
+
+					if ( ptsRelxValIter == std::end(previousIter->second) ) {
+
+						// Interpolate the relaxed value by N2_linear
+						REAL r2min_1st = std::numeric_limits<REAL>::max();
+						REAL r2min_2nd = std::numeric_limits<REAL>::max();
+						OTYPE value_1st = 0, value_2nd = 0;
+						for( size_t i = 0 ; i < previousIter->second.size() ; i++ ) {
+							REAL dr2 = normsq( focus - previousIter->second[i].first );
+							if ( dr2 < r2min_1st ) {
+								r2min_2nd = r2min_1st;
+								value_2nd = value_1st;
+								r2min_1st = dr2;
+								value_1st = previousIter->second[i].second ;
+							} else if ( dr2 < r2min_2nd ) {
+								r2min_2nd = dr2;
+								value_2nd = previousIter->second[i].second ;
+							}
+						}
+
+						REAL r1 = std::sqrt( r2min_1st );
+						REAL r2 = std::sqrt( r2min_2nd );
+
+						filteredOldValue = ( value_1st * r2 + value_2nd * r1 ) / ( r1 + r2 );
+
+					} else {
+
+						filteredOldValue = ptsRelxValIter->second;
+
+					}
+
+					std::vector<std::pair<point_type, REAL>> ptsVluTemp{
+						std::make_pair(focus,get_relaxed_value(filteredValue,filteredOldValue))
+					};
+
+					ptsTimeVlu_.push_back(std::make_pair(t, ptsVluTemp));
+
+					return get_relaxed_value(filteredValue,filteredOldValue);
+
+			} else {
+
+					auto ptsRelxValIter = std::find_if(previousIter->second.begin(), 
+						previousIter->second.end(), [focus](std::pair<point_type, REAL> b) {
+					return normsq(focus - b.first) < std::numeric_limits<REAL>::epsilon();
+					});
+
+					if ( ptsRelxValIter == std::end(previousIter->second) ) {
+
+						// Interpolate the relaxed value by N2_linear
+						REAL r2min_1st = std::numeric_limits<REAL>::max();
+						REAL r2min_2nd = std::numeric_limits<REAL>::max();
+						OTYPE value_1st = 0, value_2nd = 0;
+						for( size_t i = 0 ; i < previousIter->second.size() ; i++ ) {
+							REAL dr2 = normsq( focus - previousIter->second[i].first );
+							if ( dr2 < r2min_1st ) {
+								r2min_2nd = r2min_1st;
+								value_2nd = value_1st;
+								r2min_1st = dr2;
+								value_1st = previousIter->second[i].second ;
+							} else if ( dr2 < r2min_2nd ) {
+								r2min_2nd = dr2;
+								value_2nd = previousIter->second[i].second ;
+							}
+						}
+
+						REAL r1 = std::sqrt( r2min_1st );
+						REAL r2 = std::sqrt( r2min_2nd );
+
+						filteredOldValue = ( value_1st * r2 + value_2nd * r1 ) / ( r1 + r2 );
+
+					} else {
+
+						filteredOldValue = ptsRelxValIter->second;
+
+					}
+
+					auto ptsPresentRelxValIter = std::find_if(presentIter->second.begin(), 
+						presentIter->second.end(), [focus](std::pair<point_type, REAL> b) {
+					return normsq(focus - b.first) < std::numeric_limits<REAL>::epsilon();
+					});
+
+					if ( ptsPresentRelxValIter == std::end(presentIter->second) ) {
+
+						presentIter->second.push_back(
+							std::make_pair(focus,get_relaxed_value(
+								filteredValue,filteredOldValue)
+							)
+						);
+
+					} else {
+
+						ptsPresentRelxValIter->second = get_relaxed_value(filteredValue,filteredOldValue);
+
+					}
+
+					return get_relaxed_value(filteredValue,filteredOldValue);
+
+			}
+
+		}
+	}
+
+private:
+	template<typename OTYPE>
+	OTYPE get_relaxed_value(OTYPE filteredValue, OTYPE filteredOldValue) {
+
+		return (undRelxFac_ * filteredValue) + ((1 - undRelxFac_) * filteredOldValue);
+
 	}
 
 protected:
 	REAL initUndRelxFac_;
+	REAL undRelxFac_;
 
-	std::vector<std::pair<point_type, REAL>> ptsVluPositive_;
-	std::vector<std::pair<point_type, REAL>> ptsVluNegative_;
+	mutable std::vector<std::pair<time_type,std::vector<std::pair<point_type, REAL>>>> ptsTimeVlu_;
 
-	mutable INT posNegCounter_;
 };
 
 }
